Remove recording code, use iPhone screen capture instead
🤷

All approaches I tried so far had some issue or the other. The iPhone
screen capture seems to have a decent enough quality, and none of the
downsides. The practical concern there would be if it'll be possible
to use the iOS Photos editor to crop the clip accurately (likely), and
trim it so that the audio plays back in a loop perfectly (might be
harder, but on the plus side the macOS QuickTime editor also has a
trim functionality that might allow for more precision editing).