Fix the p5 color printer
We need to convert first to the srgb space, otherwise e.g. if we start
with a color in the oklch space, colorjs' toString will return a
string that is also in oklch.

This hadn't been a problem so far since the colors so far had already
been in the SRGB space I presume (the white/gray strokes on the become
sketch).