Link the RSS feed from the main page
It is not clear to me if the feed URL should be added to all pages, or only to
the index page. From what I've understood, it should only be on the index page,
but can also be on additional pages (e.g. /all), but it is not required to be in
the head of every page.

Refs:
https://www.rssboard.org/rss-autodiscovery
https://www.xul.fr/en-xml-rss.html
https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel
https://validator.w3.org/feed/about.html