Use solid colors
The autogenerated --mrmr-color-4, which is same as --mrmr-color-1 but at 70%
opacity, is transparent. This works fine, however, the browser is drawing the
Devanagri characters by overlaying glyphs, and so there are regions of the "a"
in "rag" (in Devanagri) which overlay each other and thus have a stronger, more
opaque color. By explicitly specifying an equivalent color in the frontmatter,
we avoid this issue.

Also, the mapping is now more consistent with the expectation of --mrmr-color-2
as a stronger version of the text color, and --mrmr-color-3 as the secondary
one.