Use the same monospace font across browsers
The default monospace on Chrome and Safari are quite different, and for the same
font-size they look quite different. This is a problem that effects our site in
general too, the default system-ui is, as the name suggests, system dependent,
but with monospace fonts we have the luxury that they're usually not in the
first fold, so we can use a network request with less concern about amateur
looking swaps.

With a fixed fixed-width font (do I like repeating words, maybe), we no longer
have to pick a font-size that works for everything and nothing, and can use a
size that should look the same everywhere.

Steps:

1. yarn add @fontsource/fira-mono
2. Add import '@fontsource/fira-mono' to gatsby-browser
3. Set that as the default code font