Thoughts on the precipice
Boundless are the expanses of space, endless the expanses of time; with my own meagre five-foot body, I attempt to fathom this vastness. But ultimately, what authority does Horatio's philosophy merit? The true nature of all that is I sum up in but a single word, and that is "unfathomable" (fukakai). Tormented by this realization, my agony has at last brought me to the decision to die. Now that I stand at the precipice, there is no anxiety of any kind in my breast. Only now do I come to understand that great tragedy and great joy are one and the same.1

How immense the universe is!
How eternal history is!
I wanted to measure the immensity with this puny five-foot body.
What authority has Horatio's philosophy?
The true nature of the whole creation.
Is in one word – “unfathomable”.
With this regret, I am determined to die.
Standing on a rock on the top of a waterfall.
I have no anxiety.
I recognize for the first time.
Great pessimism is nothing but great optimism.