https://annehelen.substack.com/p/the-diminishing-returns-of-calendar
And yes: that’s certainly how a lot of capitalist cultures think about time — as something that can be wasted or optimized. It’s often predicated on the idea that you should be focused on doing one thing, and one thing only, very efficiently: time is money, etc. etc. But that itself, sometimes referred to as a “monochronic” understanding of time, is no more or less “natural” than other ways of conceiving of time, like “polychronic” culture, which understands time as dynamic, flexible, and filled with several tasks at once, each of which will take the time that they need. Monochronic cultures may be more “efficient” in their use of time, but in their treatment of time as a commodity, they lose the richness that comes with allowing tasks, conversations, and interactions to move forward at a more natural and sustainable pace