Ernest Hemingway, Death in the Afternoon
If a writer of prose knows enough about what he is writing about he may omit things that he knows and the reader, if the writer is writing truly enough, will have a feeling of those things as strongly as though the writer had stated them. The dignity of movement of an ice-berg is due to only one-eighth of it being above water. A writer who omits things because he does not know them only makes hollow places in his writing. A writer who appreciates the seriousness of writing so little that he is anxious to make people see he is formally educated, cultured or well-bred is merely a popinjay. And this too remember; a serious writer is not to be confounded with a solemn writer. A serious writer may be a hawk or a buzzard or even a popinjay, but a solemn writer is always a bloody owl.”