Life is a very very weird machine
We always want to use our best machines to model life– it used to be mills or clockwork, now it’s computers. These turn out to be bad metaphors, especially if you go down the “genes = programs” path. Life is a very very weird machine. It has no central program or processor, it makes use of outrageously complicated mechanisms, it’s full of redundant pathways and multiple ways to do things, and it has to be understood on all levels at once. Or if you really want to simplify, the building block of life is not DNA, but the cell. But cells are not alone; they are part of tissues and a good deal of what they do is intercommunication (with each other and with the environment).

https://zompist.wordpress.com/2024/08/29/how-life-works/