J Derrida
https://philosophy.stackexchange.com/a/122172

"He didn't think there was a single, true meaning to anything. No "this is the truth"
moment with a text. For him, meaning was more like a moving target. He thought that what
a text means depends on how you read it, the ideas you bring to it, your own biases. It's
like looking at a painting. One person sees sadness, another sees beauty, another sees a
weird blob. It's not just in the painting, it's in how we experience it. He also thought
words get their meaning by being different from other words. Like, "hot" only means something
because it's not "cold". Meaning is always bouncing around, it's never stuck in one place. It's
always moving, always pointing somewhere else. And in this process there is always an element of
"deferral", that is, the meainng is always deferred to the next word, and the next and so on.

He wasn't saying "anything goes" though. It's easy to get him wrong there. He wasn't saying all
interpretations are equal or that you can just make stuff up. He was saying that the text doesn't
come with a a set meaning, and that we need to be critical of how we read it.